+++ scripts/helpers/civitai_utils.py
@@ -1,8 +1,10 @@
 import requests
 import json
+import os
+import io
+import hashlib
 from fake_useragent import UserAgent
-
-from scripts.helpers.utils import gen_sha256
+from multiprocessing import cpu_count, Process, ProcessError, Queue
 
 from modules.shared import cmd_opts, opts
 
@@ -11,6 +13,11 @@ api_urls = {
     "model_hash": "https://civitai.com/api/v1/model-versions/by-hash/",
     "model_id": "https://civitai.com/api/v1/models/",
 }
+
+
+def multithreaded_query(file_paths, query="model_hash"):
+    files_per_thread = len(file_paths) // cpu_count()
+
 
 
 def query_model_tags(file_path):
@@ -85,4 +92,62 @@ def request_civit_api(api_url=None):
             return timeout
 
     return data
+
+
+def gen_sha256(file_path):
+    """
+    Copied from the Civitai Browser+ extension
+    https://github.com/BlafKing/sd-civitai-browser-plus
+    :param file_path: full file path for model to be hashed
+    :return: sha_256 hash as str
+    """
+    json_file = os.path.splitext(file_path)[0] + ".json"
+
+    if os.path.exists(json_file):
+        try:
+            with open(json_file, 'r', encoding="utf-8") as f:
+                data = json.load(f)
+
+            if 'sha256' in data and data['sha256']:
+                hash_value = data['sha256']
+                return hash_value
+        except Exception as e:
+            print(f"SD Lora Tagger: Failed to open {json_file}; {e}")
+
+    def read_chunks(file, size=io.DEFAULT_BUFFER_SIZE):
+        while True:
+            chunk = file.read(size)
+            if not chunk:
+                break
+            yield chunk
+
+    blocksize = 1 << 20
+    h = hashlib.sha256()
+    length = 0
+    with open(os.path.realpath(file_path), 'rb') as f:
+        for block in read_chunks(f, size=blocksize):
+            length += len(block)
+            h.update(block)
+
+    hash_value = h.hexdigest()
+
+    if os.path.exists(json_file):
+        try:
+            with open(json_file, 'r', encoding="utf-8") as f:
+                data = json.load(f)
+
+            if 'sha256' in data and data['sha256']:
+                data['sha256'] = hash_value
+
+            with open(json_file, 'w', encoding="utf-8") as f:
+                json.dump(data, f, indent=4)
+        except Exception as e:
+            print(f"Failed to open {json_file}: {e}")
+    else:
+        data = {'sha256': hash_value}
+        with open(json_file, 'w', encoding="utf-8") as f:
+            json.dump(data, f, indent=4)
+
+    return hash_value
+
